pyproject.toml
[tool.poetry]
name = "smart-tg-bot"
version = "0.1.0"
description = ""
authors = ["Fiodar"]
readme = "README.md"

[tool.poetry.dependencies]
python = "^3.10"
aiogram = "^3.6.0"
requests = "^2.32.3"
google-generativeai = "^0.6.0"
python-dotenv = "^1.0.1"
async-timeout = "^4.0.3"


[build-system]
requires = ["poetry-core"]
build-backend = "poetry.core.masonry.api"
